FEAT Timer implementation
Implemented timer so that it works when user has selected their game time length of choice.

Created a new class called TimerObserver which is an interface that is implemented by the classes that need the timer. This is so that the method timerStart() can be notified when the timer instance has been initated.
As the scenes get initaited when the program is first run putting timerStart() on a class initialize() would not work as the timer instance would not have been created yet. Hence this obersever pattern is used.