fix: Fixed reset for bookshelf/Notes
Notes now clears properly after every playthrough.
Ensured that the inital commander text shows on new playthrough

Fixed the bookshelf so that the important book is always bolded.
When the game restarts the user will always see the bookshelf - no book will be pre selceted as it could be selected from the previous game.